+++ NEWS.md
@@ -1,3 +1,8 @@
+# cdmTools 1.0.5
+* Added - `RDINA2GDINA` function to increase compatibility with the `GDINA` package
+* Changed - `CA.MI` and `personFit` have been updated to be compatible with `RDINA` objects
+* Kevin Santos is included as a contributor
+
 # cdmTools 1.0.4
 * Changed - `is.Qid` has been updated to include new identifiability criteria
 * Added - `personfit` function to identify aberrant response patterns
@@ -8,7 +13,7 @@
 
 # cdmTools 1.0.2
 * Added - `CA.MI` function to calculate corrected classification accuracy with multiple imputation
-* Added - Rodrigo S. Kreitchmann as a contributor
+* Rodrigo S. Kreitchmann is included as a contributor
 
 # cdmTools 1.0.1
 * Added - `GNPC` function to estimate attribute profiles